Reorganized files in js directory
